+++ edgar-app/src/app/dashboard/chat/page.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import { VStack, Text } from '@chakra-ui/react';
 import NavBar from 'components/pages/patient/NavBar';
-import ProtectedPage from '../src/components/pages/ProtectedPage';
+import ProtectedPage from 'components/pages/ProtectedPage';
 
 const Chat = (): JSX.Element => (
 	<ProtectedPage>
