+++ edgar-app/src/app/dashboard/medical/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Box, Stack, Text, useDisclosure, VStack } from '@chakra-ui/react';
 
 import BannerCard from 'components/pages/patient/BannerCard';
@@ -14,7 +16,7 @@ import ResponsiveNavBar from 'components/pages/patient/medical/ResponsiveNavBar'
 import { usePatientContext } from 'contexts/user';
 
 import colors from 'theme/foundations/colors';
-import ProtectedPage from '../src/components/pages/ProtectedPage';
+import ProtectedPage from 'components/pages/ProtectedPage';
 
 const Medical = (): JSX.Element => {
 	const { infos } = usePatientContext();
